Fixed strange behaviour when number of alive nodes gets very low. Also, a proper wireless propagation model would be better - as is it is assumed all messages reach their destinations.

+++ nodeArchitecture.h
@@ -12,7 +12,7 @@ typedef struct Node {
 	int x;
 	int y;
 	char type;
-	int G;
+	float G;
 	float energy;
 	int clusterHead;
 	int dead;
